# internal/http/docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "HTTP API for the Codevertex Subscriptions Service — manages subscription plans, features, usage, billing, and tenant entitlements.",
        "title": "Subscriptions API",
        "contact": {},
        "version": "1.0"
    },
    "host": "pricingapi.codevertexafrica.com",
    "basePath": "/api/v1",
    "paths": {
...
        "/tenants/{tenant_id}/tokens/balance": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns the current token balance for a tenant+service. S2S (treasury-api's external eTIMS middleware) or tenant-facing.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Get API token wallet balance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant UUID",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Defaults to etims_api",
                        "name": "service_tag",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bengobox_subscription-service_internal_modules_billing.WalletSnapshot"
                        }
                    }
                }
            }
        },
        "/tenants/{tenant_id}/tokens/deduct": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Atomically deducts the resolved token cost for endpoint_pattern from the tenant's wallet, refusing (402) if the balance is insufficient. Called BEFORE the metered operation runs, not after — see treasury-api's ExternalAPIKeyAuth middleware.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Spend API tokens for one call (S2S)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant UUID",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Deduct request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers.deductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bengobox_subscription-service_internal_modules_billing.DeductResult"
                        }
                    },
                    "402": {
                        "description": "insufficient_tokens",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tenants/{tenant_id}/tokens/estimate": {
            "post": {
                "description": "Given expected call volume, computes tokens/month from the token-cost registry and compares against every published plan for service_tag, so a tenant can buy from a point of information instead of blindly.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Estimate monthly API token consumption and recommend a plan",
                "parameters": [
                    {
                        "description": "Estimate request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers.estimateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers.estimateResponse"
                        }
                    }
                }
            }
        },
        "/tenants/{tenant_id}/tokens/refund": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Reverses a Deduct when the downstream operation turned out not to have done real work (e.g. the caller's request failed validation before reaching KRA) — called by treasury-api's ExternalAPIKeyAuth middleware after the handler runs, only for status codes that indicate no real KRA attempt happened.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Refund previously-deducted API tokens (S2S)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant UUID",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Refund request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers.refundRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tenants/{tenant_id}/tokens/topup": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a Treasury payment intent for a token top-up; tokens are credited once payment is confirmed (see the api_token_topup NATS consumer branch).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Buy API tokens (self-serve)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant UUID",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Top-up request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers.topUpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tenants/{tenant_id}/tokens/transactions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscriptions"
                ],
                "summary": "List API token wallet transactions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant UUID",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Defaults to etims_api",
                        "name": "service_tag",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Default 50, max 200",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Rows to skip",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_bengobox_subscription-service_internal_modules_billing.DeductResult": {
            "type": "object",
            "properties": {
                "allowed": {
                    "type": "boolean"
                },
                "new_balance": {
                    "type": "integer"
                },
                "tokens_charged": {
                    "type": "integer"
                }
            }
        },
        "github_com_bengobox_subscription-service_internal_modules_billing.WalletSnapshot": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer"
                },
                "lifetime_granted": {
                    "type": "integer"
                },
                "low_balance": {
                    "type": "boolean"
                },
                "low_balance_threshold": {
                    "type": "integer"
                }
            }
        },
...
        "internal_http_handlers.deductRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "endpoint_pattern": {
                    "type": "string"
                },
                "ref_id": {
                    "type": "string"
                },
                "ref_type": {
                    "type": "string"
                },
                "service_tag": {
                    "type": "string"
                }
            }
        },
...
        "internal_http_handlers.estimateRequest": {
            "type": "object",
            "properties": {
                "avg_sales_per_day": {
                    "description": "Shortcut: assumes every call is a sales-transmission-weight call (the common \"N sales/day\" case).",
                    "type": "integer"
                },
                "calls_per_day": {
                    "description": "endpoint_pattern -\u003e calls/day",
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "days_per_month": {
                    "type": "integer"
                },
                "service_tag": {
                    "type": "string"
                }
            }
        },
        "internal_http_handlers.estimateResponse": {
            "type": "object",
            "properties": {
                "plans_compared": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_http_handlers.planComparison"
                    }
                },
                "recommended_plan": {
                    "type": "string"
                },
                "tokens_per_month": {
                    "type": "integer"
                }
            }
        },
...
        "internal_http_handlers.planComparison": {
            "type": "object",
            "properties": {
                "covers_estimate": {
                    "type": "boolean"
                },
                "included_tokens": {
                    "type": "integer"
                },
                "monthly_price_kes": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "plan_code": {
                    "type": "string"
                },
                "projected_topup_kes": {
                    "type": "number"
                },
                "token_price_kes": {
                    "type": "number"
                }
            }
        },
...
        "internal_http_handlers.refundRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "ref_id": {
                    "type": "string"
                },
                "service_tag": {
                    "type": "string"
                },
                "tokens": {
                    "type": "integer"
                }
            }
        },
...
        "internal_http_handlers.topUpRequest": {
            "type": "object",
            "properties": {
                "amount_kes": {
                    "type": "number"
                },
                "return_url": {
                    "type": "string"
                },
                "service_tag": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Service-to-service API key",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "JWT Bearer token from SSO (e.g. \"Bearer eyJ...\")",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}